Moved bucket client pkg
