+++ pkg/cortex/cortex_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/cortexproject/cortex/pkg/ring"
 	"github.com/cortexproject/cortex/pkg/ring/kv"
 	"github.com/cortexproject/cortex/pkg/ruler"
+	"github.com/cortexproject/cortex/pkg/storage/backend"
 	"github.com/cortexproject/cortex/pkg/storage/backend/s3"
 	"github.com/cortexproject/cortex/pkg/storage/tsdb"
 	"github.com/cortexproject/cortex/pkg/util/flagext"
@@ -28,8 +29,8 @@ func TestCortex(t *testing.T) {
 		},
 		Ingester: ingester.Config{
 			BlocksStorageConfig: tsdb.BlocksStorageConfig{
-				Bucket: tsdb.BucketConfig{
-					Backend: tsdb.BackendS3,
+				Bucket: backend.BucketConfig{
+					Backend: backend.S3,
 					S3: s3.Config{
 						Endpoint: "localhost",
 					},
@@ -46,8 +47,8 @@ func TestCortex(t *testing.T) {
 			},
 		},
 		BlocksStorage: tsdb.BlocksStorageConfig{
-			Bucket: tsdb.BucketConfig{
-				Backend: tsdb.BackendS3,
+			Bucket: backend.BucketConfig{
+				Backend: backend.S3,
 				S3: s3.Config{
 					Endpoint: "localhost",
 				},
