+++ pkg/storage/backend/bucket_client_mock.go
@@ -1,4 +1,4 @@
-package tsdb
+package backend
 
 import (
 	"bytes"
