+++ pkg/storage/backend/user_bucket_client.go
@@ -1,4 +1,4 @@
-package tsdb
+package backend
 
 import (
 	"context"
